Move obsolete resampling code.
