+++ eniric/obsolete/nIR_run.py
@@ -9,8 +9,7 @@ from typing import List, Optional, Sequence, Union
 
 import eniric
 from eniric.obsolete.nIRanalysis import convolve_spectra
-from eniric.obsolete.utilities import get_spectrum_name
-from eniric.resample import resampler
+from eniric.obsolete.utilities import get_spectrum_name, resampler
 from eniric.utilities import resolutions2ints
