+++ eniric/obsolete/utilities.py
@@ -1,14 +1,18 @@
 import os
 import re
+from genericpath import isfile
+from os.path import join
 from typing import Optional, Tuple, Union
 
 import numpy as np
 from astropy import constants as const
 from numpy.core.multiarray import ndarray
 
+import eniric
 import eniric.obsolete.IOmodule
 from eniric import IOmodule as io
 from eniric.atmosphere import consecutive_truths
+from eniric.resample import log_resample
 
 
 def get_spectrum_name(
@@ -211,3 +215,81 @@ def read_spectrum(spec_name: str) -> Tuple[ndarray, ndarray]:
         flux_photons = flux * wav_micron  # Convert to photons
 
     return wav_micron, flux_photons
+
+
+results_dir = eniric.paths["results"]
+resampled_dir = eniric.paths["resampled"]
+
+
+def resample_allfiles(
+    results_dir: Optional[str] = None, resampled_dir: Optional[str] = None
+) -> int:
+    """Resample all files inside results_dir folder.
+
+    Parameters
+    ----------
+    results_dir: str
+        Directory containing results to resample.
+    resampled_dir: str
+        Directory to save resampled results.
+    """
+    if results_dir is None:
+        results_dir = eniric.paths["results"]  # type: ignore
+    if resampled_dir is None:
+        resampled_dir = eniric.paths["resampled"]  # type: ignore
+    # Getting a list of all the files
+    onlyfiles = [f for f in os.listdir(results_dir) if isfile(join(results_dir, f))]
+
+    [
+        resampler(spectrum_file, results_dir=results_dir, resampled_dir=resampled_dir)
+        for spectrum_file in onlyfiles
+        if spectrum_file.endswith(".txt")
+    ]
+
+    return 0
+
+
+def resampler(
+    spectrum_name: str = "Spectrum_M0-PHOENIX-ACES_Yband_vsini1.0_R60k.txt",
+    results_dir: str = results_dir,
+    resampled_dir: str = resampled_dir,
+    sampling: Union[int, float] = 3.0,
+) -> int:
+    """Resample a spectrum file by interpolation onto a grid with a
+    sampling of 3 pixels per resolution element.
+
+    Parameters
+    ----------
+    spectrum_name: str
+        Name of spectrum.
+    results_dir: str
+        Directory to find the spectrum to load.
+    resample_dir: str
+        Directory to save the results.
+    sampling: float (default=3.0)
+        Sampling per pixel.
+    """
+    os.makedirs(resampled_dir, exist_ok=True)
+    read_name = os.path.join(results_dir, spectrum_name)
+
+    match = re.search("_R(\d{2,3})k", spectrum_name)
+    if match:
+        resolution = int(match.group(1)) * 1000  # type: int
+    else:
+        raise Exception("Did not match Resolution")
+
+    wavelength, __, spectrum = io.pdread_3col(read_name, noheader=True)
+    wav_grid = log_resample(wavelength, sampling, resolution)
+
+    interpolated_flux = np.interp(wav_grid, wavelength, spectrum)
+
+    output_path = [
+        resampled_dir,
+        "{0}_res{1:3.01f}.txt".format(spectrum_name[:-4], float(sampling)),
+    ]
+    filetowrite = os.path.join(*output_path)
+    eniric.obsolete.IOmodule.write_e_2col(
+        filetowrite, wav_grid[1:-2], interpolated_flux[1:-2]
+    )  # [1:-2] for border effects
+
+    return 0
